! move more items back to abstract, let themes decide what they want to override.  Only getSettings is required as a start.
